+++ Gemfile
@@ -18,6 +18,7 @@ gem "mini_magick", "4.9.5"
 gem "puma", "~> 5.0"
 gem "rails", "~> 6.1.4", ">= 6.1.4.1"
 gem "rails-i18n"
+gem "ransack"
 gem "sass-rails", ">= 6"
 gem "sqlite3", "~> 1.4"
 gem "turbolinks", "~> 5"
