+++ app/views/layouts/_sort.html.erb
@@ -1,0 +1,6 @@
+<div class="btn-group" role="group" aria-label="Basic outlined example">
+  <%= sort_link @q, :name, {}, class: "btn btn-outline-secondary" %>
+  <%= sort_link @q, :type_room, {}, class: "btn btn-outline-secondary" %>
+  <%= sort_link @q, :hourly_price, {}, class: "btn btn-outline-secondary" %>
+  <%= sort_link @q, :created_at, {}, class: "btn btn-outline-secondary" %>
+</div>
